Move CLI::Args into its own file
